Added Mockup.bmpr and Mockup-July 8.pdf
Mockup.bmpr is created using Balsamiq Mockups 3 and could be opened with that software, 
while the Mockup-July 8.pdf contains a PDF export of the mockup.

Note: A mockup is meant to describe what the app would look like, emphasizing on the functions of the app (not its aesthetic yet). It should also describe the flow of the app, i.e. how the user would be interacting with it.