created using Balsamiq Mockups 3
Mockup.bmpr is created using Balsamiq Mockups 3 and could be opened with that software

Note: A mockup is meant to describe what the app would look like, emphasizing on the functions of the app (not its aesthetic yet). It should also describe the flow of the app, i.e. how the user would be interacting with it.